+++ src/system/parse_basis.jl
@@ -1,18 +1,27 @@
 const BASIS_LIB = joinpath(@__DIR__, "..", "..", "lib", "basis_sets")
 
 """
-    parse_basis(basis_name::String, atom::Atom) 
+    parse_basis(basis_name::String, atom::Atom; fallback=false) 
 
   Search and parse the basis set for a given atom.
+  If fallback is `true`, use `def2-universal-jkfit` as a fallback basis set.
 
   Return a list of angular shells [`AngularShell`](@ref).
 """
-function parse_basis(basis_name::String, atom::Atom)
+function parse_basis(basis_name::String, atom::Atom; fallback=false)
   if startswith(basis_name, "{")
     # the basis block is given explicitly, parse basis set from string
     return parse_basis_block(strip(basis_name, ['{','}'] ) , atom)
   else
-    basisfile = basis_file(basis_name)
+    basisfile = basis_file(basis_name) 
+    if basisfile == ""
+      if fallback
+        println(atomic_center_symbol(atom),": Basis set $basis_name not found, using def2-universal-jkfit as a fallback.")
+        basisfile = basis_file("def2-universal-jkfit")
+      else
+        error("Basis set $basis_name not found!")
+      end
+    end
     basisblock = read_basis_block(basisfile, atom)
   end
   return parse_basis_block(basisblock, atom)
@@ -38,7 +47,8 @@ function basis_file(basis_name::AbstractString)
     end
   end 
   if version < 0
-    error("Basis set $basis_name not found!")
+    # Basis set not found
+    return ""
   end
   filename = "$mainname.$version.mpro"
   if !isfile(filename)